requirements.txt
# --- The Foundation (Strict Versions) ---
numpy==1.26.4
scipy==1.13.1
pandas==2.2.3
PyYAML==6.0.1
huggingface-hub==0.24.7

# --- The App Framework ---
streamlit
python-dotenv

# --- LangChain Ecosystem ---
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
langchain-chroma
sentence-transformers

# --- PDF & OCR Tools (Lightweight) ---
unstructured
pdf2image
pdfminer.six
pikepdf
pypdf
openpyxl
